kernel: Update kernel 5.4 configuration
This adds many new kernel 5.4 configuration symbols to the generic
config 5.4.

When adding a new target I was asked for an option to these options. Set
them in the generic configuration instead of doing this in each target
configuration.

The configuration was refreshed like this:
/scripts/kconfig.pl '+' target/linux/generic/config-5.4 /dev/null > target/linux/generic/config-5.4-new
mv target/linux/generic/config-5.4-new target/linux/generic/config-5.4

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>