mac80211: ath11k_nss: fix bad upload speed
With wifi offload, some buf settings are needed for NSS to
correctly work. Currently without these option upload speed
are not optimal with a result of peak speed that decrease to
a joke value after time of traffix. This is only limited to
upload speed.

Add these additional settings in a hotplug script to fix this
upload error. With this applied correct performance are observed.

With 1G profile peak traffic of 1.8gbps is reported by iperf.
With 512MB profile the correct traffic of 1.2gbps is reported.

Both results are with Intel AX210 card at 160Mhz

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>