ipq807x: rework Xiaomi AX3600 and Redmi AX6 partition layout
Rework partition logic to expand the rootfs.
- Drop dual partition support
- Enforce single partition and load always the first
  partition
- Expand the roots from 10mb or less to 200+mb and
  use full nand

To make this change we had to drop smem partition layout
and use fixed partition layout by providing our own.
Due to Xiaomi bootloader limitation, some hacks are needed
to make this work correctly.
The hardcoded bootmiwifi command does tons of change to the
installed system. One of them is putting in the partitions
node the smem partition by overwriting them. This was found
by Robi and was ""fixed"" by providing a dummy partitions node.

Another problem comes from applying an ubi layout that is different
than the one uboot expects, because of this when the command is
called, the rootfs_data is resized to the smem partition size
instead of the one provided by the kernel. This makes effectively
any fixed partition tweak useless since uboot will revert it anyway in
one or another way,

To fix this additional problem we decided to implement a new logic.
We put the kernel in a dedicated ubi partition that is the rootfs
partition uboot expect the entire system to be placed on (35mb).
Then we put rootfs and rootfs_data to the rest of the nand.
This effectively solve all the problems and permit a correct expansion
of the nand by bypassing all the uboot limitation, at the expense
of some mb wasted.

To install openwrt on this new partition layout, UART access is mandatory
but this is to prevent any sort of error where an user would cause
a brick on rootfs getting filled. (On an usable system with normal
package, you would get 9 or 8 mb of free space and that is with basic
package... adding additional stuff and your rootfs is full)

To install it you have 2 different way:
- load a initramfs image (once you have enabled uboot serial access)
  using a tftp server
  (example command)
  tftpboot 50000000 192.168.1.10:ipq807x/generic/openwrt-ipq807x-generic-xiaomi_ax3600-initramfs-fit-uImage.itb; bootm

- use the artifact nand-initramfs-factory.ubi and flash the bin from the
  running system using the following steps: (thanks dimfish for the
  following instructions)
  1. Upload the ubi image to /tmp on the router
  2. Find active rootfs
     For stock fw: nvram get flag_last_success
     For openwrt fw: nvram get flag_last_success
  3.1 If point 2 result is "0"
      ubiformat /dev/mtd13 -f /tmp/openwrt-ipq807x-generic-xiaomi_ax3600-nand-initramfs-factory.ubi
      For stock fw:
        nvram set flag_last_success=1
        nvram set flag_boot_rootfs=1
        nvram commit
        reboot
      For openwrt fw:
        fw_setenv flag_last_success 1
        fw_setenv flag_boot_rootfs 1
        reboot
  3.2 If point 2 result is "1"
      ubiformat /dev/mtd12 -f /tmp/openwrt-ipq807x-generic-xiaomi_ax3600-nand-initramfs-factory.ubi
      For stock fw:
        nvram set flag_last_success=0
        nvram set flag_boot_rootfs=0
        nvram commit
        reboot
      For openwrt fw:
        fw_setenv flag_last_success 0
        fw_setenv flag_boot_rootfs 0
        reboot
  4. With the router rebooted check that the new partition layout is
     used by the flashed firmware

      cat /proc/mtd
      dev:    size   erasesize  name
      mtd0: 00100000 00020000 "0:sbl1"
      mtd1: 00100000 00020000 "0:mibib"
      mtd2: 00300000 00020000 "0:qsee"
      mtd3: 00080000 00020000 "0:devcfg"
      mtd4: 00080000 00020000 "0:rpm"
      mtd5: 00080000 00020000 "0:cdt"
      mtd6: 00080000 00020000 "0:appsblenv"
      mtd7: 00100000 00020000 "0:appsbl"
      mtd8: 00080000 00020000 "0:art"
      mtd9: 00080000 00020000 "bdata"
      mtd10: 00100000 00020000 "pstore"
      mtd11: 023c0000 00020000 "kernel"
      mtd12: 0d240000 00020000 "rootfs"

Once the system has loaded simply load the 2 .ubi factory files with SCP and run
the following command:
ubiformat /dev/mtd11 -f /tmp/openwrt-ipq807x-generic-xiaomi_ax3600-squashfs-nand-factory-kernel.ubi
ubiformat /dev/mtd12 -f /tmp/openwrt-ipq807x-generic-xiaomi_ax3600-squashfs-nand-factory-rootfs.ubi
fw_setenv flag_boot_rootfs 0
fw_setenv flag_last_success 0
fw_setenv flag_boot_success 0
fw_setenv flag_try_sys1_failed 0
fw_setenv flag_try_sys2_failed 1

Now you can reboot the router and system will be loaded
with the new partition layout.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>